+++ .pre-commit-config.yaml
@@ -46,6 +46,8 @@ repos:
     hooks:
     -   id: mypy
         args: [--follow-imports=silent, --ignore-missing-imports,  --allow-redefinition]
+        additional_dependencies: [types-requests]
+
 -   repo: https://github.com/pycqa/pydocstyle
     rev: 6.1.1
     hooks:
