+++ Migrators/semmed/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Define functions for migrating data from the SemMed database to TypeDB."""
